+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ConstrainSolver.jl - Changelog
 
+## Unreleased
+- CS.Integers
+  - i.e `@variable(m, x, CS.Integers([1,2,5]))`
+- **Bugfixes:**
+  - return infeasible if start fixing is infeasible [#132](https://github.com/Wikunia/ConstraintSolver.jl/pull/132) 
+
 ## 0.1.0 (7th of April 2020)
 Initial implementation of the solver:
 Basic constraints: